requirements.txt
django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.6,<5.0
django-filter>=24.3,<25.0
dj-database-url>=2.3,<3.0
psycopg[binary]>=3.2,<4.0
whitenoise>=6.8,<7.0
gunicorn>=23.0,<24.0
openai>=1.58,<2.0
python-dotenv>=1.0,<2.0
